requirements.txt
beautifulsoup4==4.12.3
openpyxl==3.1.2
selenium==4.20.0
webdriver-manager==4.0.2
requests==2.31.0
sseclient-py==0.0.28
groq>=0.25.0
pandas>=2.0.0
python-dateutil>=2.8.2
lxml>=4.9.3
colorama>=0.4.6
tqdm>=4.66.0
pytz>=2023.3